+++ pyproject.toml
@@ -48,6 +48,8 @@ dependencies = [
   "cryptography==46.0.7",
   "regex==2026.2.28",
   "pythonnet==3.1.0-rc0 ; sys_platform == 'win32'",
+  "pygls>=2.0.0",
+  "lsprotocol>=2023.0.0",
 ]
 
 [tool.uv.sources]
